221025 25635 complete commit

+++ 3.Gold/20000-29999/25635/25635.cpp
@@ -14,16 +14,16 @@
 
 using namespace std;
 
-int play[100000];
+long long play[100000];
 
 int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
     cout.tie(NULL);
 
-    int n;
-    int sum = 0;
-    int maxn = 0;
+    long long n;
+    long long sum = 0;
+    long long maxn = 0;
     cin >> n;
     for(int i = 0; i < n; i++) {
         cin >> play[i];
